// projects/facility-management/src/app/shared/modules/copy-it-print-details/models/copy-it.enum.ts
/** CopyItPageType */
export enum CopyItPageType {
    BOTH = 1,
    COVER = 2,
    BACK = 3
}

/** Enum for copyIt Page */
export enum CopyItPage {
    FRONT_PAGE = 1,
    MIDDLE_PAGE = 2
}

/** Available paper color */
export enum CopyItPaperColorType {
    BW = 1,
    COLOR = 2
}

/** Available paper side */
export enum CopyItPaperSideType {
    ONE_SIDED = 1,
    TWO_SIDED = 2
}

/** to-do */
export enum RateRequestType {
    CLIENT = 1,
    TENANT = 2
}

/** CopyItProof */
export enum CopyItProof {
    YES = 1,
    NO = 2
}

/** CopyItPriceQuote */
export enum CopyItPriceQuote {
    YES = 1,
    NO = 2
}

/** CopyIt Reproduction Type */
export enum CopyItReproductionType {
    BLACK_WHITE = 1,
    COLOR = 2,
    SCAN = 3
}

/** CopyIt charge section type */
export enum CopyItCostType {
    ESTIMATES = 1,
    CHARGES = 2
}